+++ src/test/java/test/com/pmrodrigues/ellasa/repositories/TestUsuarioRepository.java
@@ -5,11 +5,15 @@ import org.junit.Assert;
 import org.junit.Before;
 import org.junit.Test;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.jdbc.core.RowMapper;
 import org.springframework.test.context.ContextConfiguration;
 import org.springframework.test.context.junit4.AbstractTransactionalJUnit4SpringContextTests;
 
 import com.pmrodrigues.ellasa.models.Usuario;
 import com.pmrodrigues.ellasa.repositories.UsuarioRepository;
+
+import java.sql.ResultSet;
+import java.sql.SQLException;
 
 @ContextConfiguration(locations = {"classpath:test-applicationContext.xml"})
 public class TestUsuarioRepository
@@ -19,43 +23,27 @@ public class TestUsuarioRepository
 	@Autowired
 	private UsuarioRepository repository;
 
-	@SuppressWarnings("deprecation")
 	private void deleteDadosDeTeste() {
 
-		if (this.jdbcTemplate.queryForInt(
-				"select count(1) from usuario where email = ?",
-				"marsilvarodrigues@gmail.com") > 0) {
-			final Long id = this.jdbcTemplate.queryForLong(
-					"select id from usuario where email = ?",
-					"marsilvarodrigues@gmail.com");
+        this.jdbcTemplate.query("select id , residencial_id , celular_id from usuario where email = 'marsilvarodrigues@gmail.com'" , new RowMapper<Object>() {
+            @Override
+            public Object mapRow(ResultSet rs, int rowNum) throws SQLException {
 
-			final Long celular_id = this.jdbcTemplate.queryForLong(
-					"select celular_id from usuario where id = ?", id);
-			final Long residencial_id = this.jdbcTemplate.queryForLong(
-					"select residencial_id from usuario where id = ?", id);
+                final Long userId = rs.getLong("id");
+                final Long celularId = rs.getLong("celular_id");
+                final Long residencialId = rs.getLong("residencial_id");
 
-			if (this.jdbcTemplate.queryForInt(
-					"select count(id) from contrato where franqueado_id = ?",
-					id) > 0) {
-				Long contrato_id = this.jdbcTemplate.queryForLong(
-						"select id from contrato where franqueado_id = ?", id);
+                TestUsuarioRepository.this.jdbcTemplate.update("delete from ordempagamento where contrato_id = ( select id from contrato where franqueado_id = ?)" , userId);
+                TestUsuarioRepository.this.jdbcTemplate.update("delete from contrato where franqueado_id = ?" , userId);
+                TestUsuarioRepository.this.jdbcTemplate.update("delete from franqueadopessoafisica where id = ?", userId);
+                TestUsuarioRepository.this.jdbcTemplate.update("delete from franqueadopessoajuridica where id = ?", userId);
+                TestUsuarioRepository.this.jdbcTemplate.update("delete from franqueado where id = ?", userId);
+                TestUsuarioRepository.this.jdbcTemplate.update("delete from usuario where id = ?" , userId);
+                TestUsuarioRepository.this.jdbcTemplate.update("delete from telefone where id in (?,?)" , celularId , residencialId);
 
-				this.jdbcTemplate.update(
-						"delete from ordempagamento where contrato_id = ?",
-						contrato_id);
-				this.jdbcTemplate.update(
-						"delete from contrato where franqueado_id = ?", id);
-			}
-			this.jdbcTemplate.update(
-					"delete from franqueadopessoafisica where id = ?", id);
-			this.jdbcTemplate.update(
-					"delete from franqueadopessoajuridica where id = ?", id);
-			this.jdbcTemplate.update("delete from franqueado where id = ?", id);
-			this.jdbcTemplate.update("delete from usuario where id = ?", id);
-			this.jdbcTemplate.update("delete from telefone where id in (?,?)",
-					celular_id, residencial_id);
-
-		}
+                return null;
+            };
+        });
 	}
 
 	@Before
